# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API Cửa hàng quần áo'
  description: 'Tài liệu API cho hệ thống cửa hàng quần áo'
  version: 1.0.0
paths:
...
  '/api/attributes/{attributeType}/values':
    get:
      tags:
        - 'Thuộc tính'
      summary: 'Danh sách giá trị thuộc tính'
      description: 'Có thể truyền ID hoặc name của loại thuộc tính.'
      operationId: listAttributeValues
      parameters:
        -
          name: attributeType
          in: path
          required: true
          schema:
            type: string
          example: color
      responses:
        200:
          description: 'Lấy danh sách giá trị thuộc tính thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttributeValue'
                type: object
        404:
          description: 'Không tìm thấy loại thuộc tính'
    post:
      tags:
        - 'Thuộc tính'
      summary: 'Tạo giá trị thuộc tính'
      operationId: createAttributeValue
      parameters:
        -
          name: attributeType
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - value
              properties:
                value:
                  type: string
                  example: black
                display_value:
                  description: 'Mặc định bằng value nếu không gửi.'
                  type: string
                  example: Đen
                  nullable: true
                meta_data:
                  type: object
                  example:
                    hex: '#000000'
                  nullable: true
              type: object
      responses:
        201:
          description: 'Tạo giá trị thuộc tính thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 201
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    $ref: '#/components/schemas/AttributeValue'
                type: object
        401:
          description: 'Chưa xác thực'
        403:
          description: 'Chỉ quản trị viên được phép'
        404:
          description: 'Không tìm thấy loại thuộc tính'
        422:
          description: 'Dữ liệu không hợp lệ hoặc value đã tồn tại'
      security:
        -
          bearerAuth: []
  '/api/attributes/{attributeType}/values/{attributeValue}':
    get:
      tags:
        - 'Thuộc tính'
      summary: 'Chi tiết giá trị thuộc tính'
      operationId: getAttributeValue
      parameters:
        -
          name: attributeType
          in: path
          required: true
          schema:
            type: string
          example: color
        -
          name: attributeValue
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: 'Lấy giá trị thuộc tính thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    $ref: '#/components/schemas/AttributeValue'
                type: object
        404:
          description: 'Không tìm thấy loại hoặc giá trị thuộc tính'
    put:
      tags:
        - 'Thuộc tính'
      summary: 'Cập nhật giá trị thuộc tính'
      operationId: updateAttributeValue
      parameters:
        -
          name: attributeType
          in: path
          required: true
          schema:
            type: integer
          example: 1
        -
          name: attributeValue
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                value:
                  type: string
                  example: black
                display_value:
                  type: string
                  example: 'Màu đen'
                meta_data:
                  type: object
                  example:
                    hex: '#000000'
                  nullable: true
              type: object
      responses:
        200:
          description: 'Cập nhật giá trị thuộc tính thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    $ref: '#/components/schemas/AttributeValue'
                type: object
        401:
          description: 'Chưa xác thực'
        403:
          description: 'Chỉ quản trị viên được phép'
        404:
          description: 'Không tìm thấy loại hoặc giá trị thuộc tính'
        422:
          description: 'Dữ liệu không hợp lệ hoặc value đã tồn tại'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Thuộc tính'
      summary: 'Xóa giá trị thuộc tính'
      description: 'Không cho phép xóa nếu giá trị đang được gắn với bất kỳ biến thể sản phẩm nào.'
      operationId: deleteAttributeValue
      parameters:
        -
          name: attributeType
          in: path
          required: true
          schema:
            type: integer
          example: 1
        -
          name: attributeValue
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: 'Xóa giá trị thuộc tính thành công'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    type: object
                    example: null
                    nullable: true
                type: object
        401:
          description: 'Chưa xác thực'
        403:
          description: 'Chỉ quản trị viên được phép'
        404:
          description: 'Không tìm thấy loại hoặc giá trị thuộc tính'
        409:
          description: 'Giá trị thuộc tính đang được sản phẩm sử dụng'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: integer
                    example: 409
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Không thể xóa vì giá trị thuộc tính đang được sử dụng.'
                  data:
                    properties:
                      usages:
                        type: array
                        items:
                          properties:
                            product_id:
                              type: integer
                              example: 10
                            product_name:
                              type: string
                              example: 'Áo thun'
                              nullable: true
                            product_variant_id:
                              type: integer
                              example: 21
                            sku:
                              type: string
                              example: AO-DEN-M
                          type: object
                    type: object
                type: object
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    AttributeValue:
      required:
        - id
        - attribute_type_id
        - value
        - display_value
      properties:
        id:
          type: integer
          example: 1
        attribute_type_id:
          type: integer
          example: 1
        value:
          type: string
          example: black
        display_value:
          type: string
          example: Đen
        meta_data:
          type: object
          example:
            hex: '#000000'
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Nhập JWT token theo dạng Bearer token.'
      bearerFormat: JWT
      scheme: bearer
...